[TECH-496] Move the run mpyl functions out of the cli folder
branch: feature/TECH-496-improve-code-quality
